requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
boto3==1.17.109
botocore==1.20.109
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
decorator==5.0.9
distlib==0.3.1
dj-static==0.0.6
Django==3.1.7
django-filter==2.4.0
django-ordered-model==3.4.3
django-rest-knox==4.1.0
djangorestframework==3.12.4
dominate==2.6.0
dynaconf==3.1.4
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
hvac==0.10.14
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
knox==0.1.14
lazy-object-proxy==1.6.0
loguru==0.5.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pylint==2.7.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
shortuuid==1.0.1
six==1.15.0
SQLAlchemy==1.4.2
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
urllib3==1.26.6
validators==0.18.2
virtualenv==20.4.3
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
